ported r92669 from branches/1.0
Utilities registered with an absolute path were returned with the RequestContainer in the aq_chain. As the result of the first utility look-up is stored in the adapter look-up cache, subsequent utility look-ups return the utlitiy with the RequestContainer of the first look-up.
  
Solution: For utilities registered with an absolute path the RequestContainer is now also removed at look-up.